Eliminate multiple settings to extruder_duplication_enabled in G28
If the IDEX mode was previously set (via GCode or the LCD Panel) to a duplication mode...   It was lost after a G28 due to this redundant setting of 'extruder_duplication_enabled'. 

It is very important to the various use cases of IDEX machines that extruder_duplication_enabled is preserved across a G28.   The reason is this allows a person to use the LCD Panel to put the machine into a duplication mode, and then send a 'normal' part file to the printer and get two parts out of the process.